Added room number UI, fixed door spawns, made stacking mods better
For some reason the door spawning that worked until now had decided that it wouldn't work anymore because the room number was defaulted to 1 when instantiating which sorta makes sense but at the same time why did it work before ?

Room UI temporary ?  Or at least not always there I think it breaks the immersion.

Stacking mod effects is now done easily using the id attribute (name sucks but couldn't find a better one) which represent the number of mod there are (mod with id 1 = first one, mod with id 2 = second one etc...) not the best solution but hey it works.